Split-up tests for repo organisation
